renamed ownerId to userId
