+++ backend/routes/api/feeds.js
@@ -10,7 +10,7 @@ router.get('/user/:userId', asyncHandler(async (req, res) => {
 	const userId = req.params.userId
 	const feeds = await Feed.findAll(
 		{
-			where: { ownerId: userId },
+			where: { userId: userId },
 			include: Source
 		}
 	);
@@ -37,9 +37,9 @@ router.post('/new', asyncHandler(async function(req, res) {
 }));
 
 router.put('/update/:id', asyncHandler(async function(req, res) {
-	const { ownerId, feedName } = req.body
+	const { userId, feedName } = req.body
 	const feed = await Feed.findByPk(req.params.id);
-	await feed.update({ ownerId, feedName });
+	await feed.update({ userId, feedName });
 	return res.json(comments);
 }));
