+++ frontend/src/components/SideBar/index.js
@@ -5,7 +5,8 @@ import { useDispatch, useSelector } from 'react-redux';
 import { getFeeds, addFeed } from '../../store/feeds';
 
 // components
-import FeedsContainer from './Feeds'
+import SingleFeed from './Feed'
+import SourceContainer from './Sources'
 
 import './side_bar.css';
 
@@ -28,8 +29,8 @@ function SideBar({ isLoaded }){
 	const handleSubmit = async (e) => {
 		e.preventDefault();
 		const data = {
-      // ownerId:sessionUser.id,
-			ownerId:1,
+      // userId:sessionUser.id,
+			userId:1,
 			name:feedName
 		}
 		let createdFeed = await dispatch(addFeed(data))
@@ -41,7 +42,20 @@ function SideBar({ isLoaded }){
       <div>Read Later</div>
       <div>Feeds</div>
       <div>All</div>
-      <FeedsContainer feeds={feeds}/>
+
+      <div className='feeds_container'>
+      {feeds && feeds?.map( feed => (
+          <div className='feed_container' key={`feed${feed?.id}`}>
+      
+            <SingleFeed feed={feed}/>
+
+            <div className='source_container'>
+              <SourceContainer sources={feed.Sources} />
+            </div>
+          </div>
+        ))}
+      </div>
+
       <div onClick={() => {setShowAddFeed(!showAddFeed)}}>Create New Feed</div>
       {showAddFeed && (
         <form onSubmit={handleSubmit} className='add_feed_form'>
